# catalog-service/docs/swagger.yaml
definitions:
  dto.AddBookRequest:
    properties:
      authorId:
        minimum: 1
        type: integer
      category:
        type: string
      pages:
        items:
          $ref: '#/definitions/dto.CreatePageRequest'
        type: array
      title:
        type: string
      year:
        type: integer
    required:
    - authorId
    - category
    - title
    - year
    type: object
  dto.Author:
    properties:
      fullname:
        type: string
      id:
        type: integer
    type: object
  dto.Book:
    properties:
      author:
        $ref: '#/definitions/dto.Author'
      category:
        type: string
      id:
        type: integer
      title:
        type: string
      year:
        type: integer
    type: object
  dto.BookViews:
    properties:
      views:
        type: integer
    type: object
  dto.CreateAuthorRequest:
    properties:
      fullname:
        type: string
    required:
    - fullname
    type: object
  dto.CreatePageRequest:
    properties:
      content:
        type: string
      number:
        type: integer
    required:
    - content
    - number
    type: object
  dto.Error:
    properties:
      error:
        type: string
    type: object
  dto.Page:
    properties:
      content:
        type: string
      id:
        type: integer
      number:
        type: integer
    type: object
info:
  contact: {}
paths:
  /catalog/authors:
    post:
      description: Adds a new author
      parameters:
      - description: Author info
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAuthorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Author'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: The token is missing, invalid or expired
          schema:
            $ref: '#/definitions/dto.Error'
        "403":
          description: The token is valid, but lacks permission
          schema:
            $ref: '#/definitions/dto.Error'
        "409":
          description: Entity already exists
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - BearerAuth: []
      summary: Create a new author
      tags:
      - catalog
  /catalog/authors/{authorID}:
    delete:
      description: Deletes an author by ID
      parameters:
      - description: Author ID
        in: path
        name: authorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: The token is missing, invalid or expired
          schema:
            $ref: '#/definitions/dto.Error'
        "403":
          description: The token is valid, but lacks permission
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - BearerAuth: []
      summary: Delete an author
      tags:
      - catalog
  /catalog/authors/{authorID}/books:
    get:
      description: Returns all books for the given author
      parameters:
      - description: Author ID
        in: path
        name: authorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Book'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Get books by author ID
      tags:
      - catalog
  /catalog/books:
    post:
      description: Creates a new book entry
      parameters:
      - description: Book info
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.AddBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Book'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: The token is missing, invalid or expired
          schema:
            $ref: '#/definitions/dto.Error'
        "403":
          description: The token is valid, but lacks permission
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/dto.Error'
        "409":
          description: Entity already exists
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - BearerAuth: []
      summary: Add a new book
      tags:
      - catalog
  /catalog/books/{bookID}:
    delete:
      description: Deletes a book by ID
      parameters:
      - description: Book ID
        in: path
        name: bookID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: The token is missing, invalid or expired
          schema:
            $ref: '#/definitions/dto.Error'
        "403":
          description: The token is valid, but lacks permission
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - BearerAuth: []
      summary: Delete a book
      tags:
      - catalog
    get:
      description: Returns content of a specific book page
      parameters:
      - description: Book ID
        in: path
        name: bookID
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Page'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Get a book page
      tags:
      - catalog
  /catalog/books/{bookID}/preview:
    get:
      description: Returns preview information for a book
      parameters:
      - description: Book ID
        in: path
        name: bookID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Book'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Preview a book
      tags:
      - catalog
  /catalog/books/{bookID}/views:
    get:
      description: Returns total number of views by different authorized users only.
        If book doesn't exist it still will return zero views count
      parameters:
      - description: Book ID
        in: path
        name: bookID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookViews'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Get views count for a book
      tags:
      - catalog
  /catalog/books/categories:
    get:
      description: Returns a list of all available book categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Get all book categories
      tags:
      - catalog
  /catalog/books/categories/{categoryName}:
    get:
      description: Returns paginated list of books for the given category
      parameters:
      - description: Category name
        in: path
        name: categoryName
        required: true
        type: string
      - description: Page number (min=1, default=1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (min=1, max=100, default=20)
        in: query
        name: count
        type: integer
      - description: Sort field (title / year / category, default=title)
        in: query
        name: sort
        type: string
      - description: Sort order (asc / desc, default=asc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Book'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: List books by category
      tags:
      - catalog
  /catalog/books/new:
    get:
      description: Returns list of recently added books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Book'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Get new books
      tags:
      - catalog
  /catalog/books/popular:
    get:
      description: Returns list of most popular books based on views count by different
        authorized users only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Book'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Get popular books
      tags:
      - catalog
  /catalog/books/search:
    get:
      description: List books by author name and/or title with pagination
      parameters:
      - description: Author name
        in: query
        name: author
        type: string
      - description: Book title
        in: query
        name: title
        type: string
      - description: Page number (min=1, default=1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (min=1, max=100, default=20)
        in: query
        name: count
        type: integer
      - description: Sort field (title / year / category, default=title)
        in: query
        name: sort
        type: string
      - description: Sort order (asc / desc, default=asc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Book'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Error'
      summary: Search books
      tags:
      - catalog
swagger: "2.0"